Move `flip` to `tensor/subtensor.py`, add docstring
