+++ tests/tensor/test_pad.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 import pytensor
-from pytensor.tensor.pad import PadMode, flip, pad
+from pytensor.tensor.pad import PadMode, pad
 
 
 floatX = pytensor.config.floatX
@@ -222,27 +222,3 @@ def test_nd_padding(mode, padding):
     f = pytensor.function([], z, mode="FAST_COMPILE")
 
     np.testing.assert_allclose(expected, f(), atol=ATOL, rtol=RTOL)
-
-
-@pytest.mark.parametrize(
-    "size", [(3,), (3, 3), (3, 5, 5)], ids=["1d", "2d square", "3d square"]
-)
-def test_flip(size: tuple[int]):
-    from itertools import combinations
-
-    x = np.random.normal(size=size).astype(floatX)
-    x_pt = pytensor.tensor.tensor(shape=size, name="x")
-    expected = np.flip(x, axis=None)
-    z = flip(x_pt, axis=None)
-    f = pytensor.function([x_pt], z, mode="FAST_COMPILE")
-    np.testing.assert_allclose(expected, f(x), atol=ATOL, rtol=RTOL)
-
-    # Test all combinations of axes
-    flip_options = [
-        axes for i in range(1, x.ndim + 1) for axes in combinations(range(x.ndim), r=i)
-    ]
-    for axes in flip_options:
-        expected = np.flip(x, axis=list(axes))
-        z = flip(x_pt, axis=list(axes))
-        f = pytensor.function([x_pt], z, mode="FAST_COMPILE")
-        np.testing.assert_allclose(expected, f(x), atol=ATOL, rtol=RTOL)
